nemo_retriever/pyproject.toml
[build-system]
requires = ["setuptools>=78.1.1", "wheel>=0.46.2"]
build-backend = "setuptools.build_meta"

[tool.uv]
managed = true
package = true
no-build-package = ["nemotron-ocr"]
# fastparquet 2026.3.0 has no wheel for Linux x86_64 (manylinux); constrain to last version that does.
override-dependencies = [
  "fastparquet>=2024.11.0,<2026",
]

[project]
name = "nemo-retriever"
dynamic = ["version"]
description = "A modern RAG ingestion pipeline from Nvidia"
readme = "README.md"
requires-python = ">=3.12,<3.13"
authors = [
  { name = "Jeremy Dyer", email = "jdyer@nvidia.com" },
]
license = "Apache-2.0"
license-files = ["LICENSE"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Operating System :: OS Independent",
]
dependencies = [
  # Core orchestration and framework
  "ray[data,serve]>=2.49.0",
  "pandas>=2.0,<3",
  "sqlglot>=30.0.0",
  "tqdm>=4.66.0",
  "typer>=0.12.0",
  "pyyaml>=6.0",
  # Service layer
  "fastapi>=0.114.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  # MCP server for AI agent integration
  "fastmcp>=2.0.0",
  # HTTP clients
  "httpx>=0.27.0",
  "requests>=2.32.5",
  "urllib3==2.6.3",
  # Utilities
  "pydantic>=2.8.0",
  "rich>=13.7.0",
  "universal-pathlib>=0.2.0",
  "numpy>=1.26.0",
  "debugpy>=1.8.0",
  # Core ingest packages
  "nv-ingest-api",
  "nv-ingest-client",
  # Document parsing and NIM client libs
  "pypdfium2==4.30.0",
  "pillow==12.1.1",
  "nltk==3.9.3",
  "markitdown",
  "langchain-nvidia-ai-endpoints>=0.3.0",
  # Default VDB solution
  "lancedb",
]

[project.optional-dependencies]

# ── Local model inference (GPU assumed; torch resolves to CUDA on Linux) ─────
# Adds HuggingFace transformers, torch, nemotron models, GPU monitoring, and vLLM.
local = [
  "transformers>=4.57.6,<5",
  "tokenizers>=0.20.3",
  "accelerate==1.12.0",
  "torch~=2.10.0",
  "torchvision>=0.25,<0.26",
  "einops",
  "easydict",
  "addict",
  "scikit-learn>=1.6.0",
  "timm==1.0.22",
  "albumentations==2.0.8",
  "nemotron-page-elements-v3>=0.dev0",
  "nemotron-graphic-elements-v1>=0.dev0",
  "nemotron-table-structure-v1>=0.dev0",
  "nemotron-ocr>=0.dev0; sys_platform == 'linux'",
  "nvidia-ml-py",
  "apscheduler>=3.10",
  "psutil>=5.9.0",
  "vllm==0.17.0; sys_platform == 'linux'",
]

# ── Multimedia — audio/ASR and SVG rendering ────────────────────────────────
# soundfile + scipy enable local Parakeet ASR on audio/video content.
# cairosvg enables SVG-to-image rendering (requires libcairo system library).
multimedia = [
  "soundfile>=0.12.0",
  "scipy>=1.11.0",
  "cairosvg>=2.7.0",
]

# ── Independent extras ───────────────────────────────────────────────────────

# SQL and graph storage backends (lancedb is in core deps).
stores = [
  "duckdb>=1.2.0",
  "duckdb-engine>=0.13.0",
  "neo4j>=5.0",
]

# BEIR benchmarking and evaluation tools (not needed for production use).
benchmarks = [
  "datasets>=4.8.2",
  "open-clip-torch==3.2.0",
]

eval = [
  "litellm>=1.40.0",
  "pyyaml>=6.0",
  "tenacity>=8.0.0",
]
dev = [
  "build>=1.2.2",
  "pytest>=8.0.2",
]

# ── Convenience: full install ─────────────────────────────────────────────────
all = [
  "nemo_retriever[local,multimedia,stores,benchmarks]",
]

[project.scripts]
retriever = "nemo_retriever.__main__:main"
retriever-harness = "nemo_retriever.harness:main"

[tool.setuptools.dynamic]
version = {attr = "nemo_retriever.version.get_build_version"}

[tool.uv.sources]
nv-ingest-api = { path = "../api/", editable = true }
nv-ingest-client = { path = "../client/", editable = true }
nemotron-page-elements-v3 = { index = "test-pypi" }
nemotron-graphic-elements-v1 = { index = "test-pypi" }
nemotron-table-structure-v1 = { index = "test-pypi" }
nemotron-ocr = { index = "test-pypi" }
# On Linux, resolve torch/torchvision from the CUDA wheel index.
# On Mac, fall through to PyPI to get CPU wheels.
torch = [
  { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
vllm = [
  { url = "https://github.com/vllm-project/vllm/releases/download/v0.17.0/vllm-0.17.0+cu130-cp38-abi3-manylinux_2_35_x86_64.whl", marker = "platform_machine == 'x86_64'" },
  { url = "https://github.com/vllm-project/vllm/releases/download/v0.17.0/vllm-0.17.0+cu130-cp38-abi3-manylinux_2_35_aarch64.whl", marker = "platform_machine == 'aarch64'" },
]

[[tool.uv.index]]
name = "test-pypi"
url = "https://test.pypi.org/simple/"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"nemo_retriever.harness.portal" = ["static/**/*"]
"nemo_retriever.service" = ["retriever-service.yaml"]
